+++ src/index.html
@@ -81,8 +81,8 @@
           >
         </li>
         <li>
-          <calcite-link href="/demos/calcite-color-picker.html"
-            >calcite-color-picker</calcite-link
+          <calcite-link href="/demos/calcite-color.html"
+            >calcite-color</calcite-link
           >
         </li>
         <li>
